% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{create_tagger}
\alias{create_tagger}
\title{Create a tagger function}
\usage{
create_tagger(
  dictionary_path,
  config_file = system.file("resources/sudachi.json", package = "sudachir2"),
  resource_dir = system.file("resources", package = "sudachir2"),
  mode = c("C", "A", "B")
)
}
\arguments{
\item{dictionary_path}{A path to a dictionary file
such as \code{"system_core.dic"}.}

\item{config_file}{A path to a config file.}

\item{resource_dir}{A path to a resource directory.}

\item{mode}{Split mode for 'sudachi.rs'.
Either \code{"C"}, \code{"A"}, or \code{"B"}.}
}
\value{
A function inheriting class \code{purrr_function_partial}.
}
\description{
Create a tagger function
}
\details{
This function just returns a wrapper function for tokenization,
i.e. does not actually create a tagger instance.
Even if arguments are invalid, this function does not raise any errors.
}
