% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-dict.R
\name{fetch_dict}
\alias{fetch_dict}
\title{Download and unarchive a dictionary for 'Sudachi'}
\usage{
fetch_dict(
  exdir,
  dict_version = "latest",
  dict_type = c("small", "core", "full")
)
}
\arguments{
\item{exdir}{Directory where the dictionary will be unarchived.}

\item{dict_version}{Version of the dictionary to be downloaded.}

\item{dict_type}{Type of the dictionary to be downloaded.
Either \code{"small"}, \code{"core"}, or \code{"full"}.}
}
\value{
\code{exdir} is invisibly returned.
}
\description{
Download and unarchive a dictionary for 'Sudachi'
}
