% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{sudachir2-package}
\alias{sudachir2}
\alias{sudachir2-package}
\title{sudachir2: R Wrapper for 'sudachi.rs'}
\description{
Offers bindings to 'sudachi.rs' (\url{https://github.com/WorksApplications/sudachi.rs}), a Rust implementation of 'Sudachi' Japanese morphological analyzer.
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
